Adds auto-close call
After creating the WAYWOTW thread this code will make another post to
auto close the topic on the next sunday at midnight.

Fixes: #1

I'm not super happy with the way we are parsing the response to get the
topic slug and id. Go won't auto push JSON numbers into a string if
asking it for a map of string/string. A future improvement could be to
create a struct that looks for these two values in general.